fix probleme migration

+++ database/migrations/2023_08_15_092200_create_freins_table.php
@@ -13,10 +13,10 @@ return new class extends Migration
     {
         Schema::create('freins', function (Blueprint $table) {
             $table->id();
-            $table->float('prix_min')->nullable();
-            $table->float('prix_max')->nullable();
-            $table->float('superficie_min')->nullable();
-            $table->float('superficie_max')->nullable();
+            $table->double('prix_min', 12, 2)->nullable();
+            $table->double('prix_max', 12, 2)->nullable();
+            $table->double('superficie_min', 12, 2)->nullable();
+            $table->double('superficie_max', 12, 2)->nullable();
             $table->boolean('etat')->default(false)->comment('1 attent 2 existe_bien_disponible 3=>traite');
             $table->double('avance')->nullable();
             $table->boolean('tranche')->default(false);
